pyproject.toml
[tool.poetry]
name = "slack-extractor-web"
version = "0.1.0"
description = "Web application for extracting Slack messages"
authors = ["Vladimir Klimov <vladimir.klimov@hyperskill.org>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.9"
slack-sdk = "3.34.0"
python-dotenv = "1.0.1"
requests = "2.32.2"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
]
addopts = "--cov=api --cov-report=term-missing"
